fix: logout link
show the logout link in all pages
